+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Clarified that items and blocks max stack size is 64
+* Removed farmers' delight integration data pack
 
 ## 4.30.1.2
 * Hopefully fixed a crash with VS Clockwork